+++ react/FormInputUpload.tsx
@@ -1,0 +1,3 @@
+import InputUpload from './components/Upload/Upload'
+
+export default InputUpload