+++ react/components/Input.tsx
@@ -11,9 +11,9 @@ import { useFormattedError } from '../hooks/useErrorMessage'
 import { FormRawInputProps } from '../typings/InputProps'
 
 export const HiddenInput: FC<FormRawInputProps> = props => {
-  const { pointer } = props
+  const { pointer, value } = props
   const inputObject = useHidden(pointer)
-  return <input {...inputObject.getInputProps()} />
+  return <input {...inputObject.getInputProps()} value={value} />
 }
 
 export const PasswordInput: FC<FormRawInputProps> = props => {