+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Adding upload input.
+
 ## [0.4.0] - 2020-11-13
 ### Added
 - Prop `placeholder` for text and textarea inputs.